requirements.txt
python-dotenv
ipykernel
langchain
openai
langchain_groq
langchain_openai
langchain_core
fastapi
uvicorn
langserve
sse_starlette